Ignore permutations with non-matching tiers
We've previously included all permutations of local transactions and local
authorities, but some of those will never be shown to users so should be
excluded from the output. If a transaction's tiers are ['district', 'unitary']
then there shouldn't be any route for a user to get to a county council's
results page for it, so even if the county council has a link in the LDG data
we need to ignore it here.

Also include the updated output - it now has 41430 rows (26% fewer) with
23469 unique interaction URLs (3% fewer).