Add results of fetching all pages
This is a bit broken still because the script errored the first time, and the
cache checking wasn't working properly the second time, so some duplicate
records were logged. I've cleaned up some of those, but some are still here.

This also doesn't include the URLs that are added back by filtering out the X
URLs from the LDG CSV - it'll need another run to pick those up. The few URLs
which are too long to be used as filenames also need truncating before another
run.