Add unique pageview data to the list of URLs used for local transactions
This is based on GA data for the top 15,000 local_transaction format pages by
pageviews for the last 6 months (20150708-20160208). That data includes the
start pages for the local transactions as well, so only 14,852 rows have
pageview data in the output. That already gets us down to pages which have had
only 10 unique pageviews in the last 6 months, so we're a long way into the
long tail at that point.