Add list of URLs used with added data from @daibach
This is urls-used-for-local-transactions-with-statuses.csv with 4 columns
added which contain data from @daibach's local services explorer:

- id (at the start): a numeric id for each row
- jumbled_url: a URL which is identical to the interaction_url_used except that
the last part of the path has been reversed, or certain query params replaced
with timestamps
- jumbled_url_status_code: the status code returned for a request to the
jumbled_url
- jumbled_url_test_date: the date on which the jumbled_url was last tested

The local services explorer generates these jumbled URLs and requests them to
test whether sites are capable of returning a 404 status code for URLs similar
to the one we're using. If the original URL returned 200 and the jumbled one
404, that's a good indication that the original 200 was really a useful
content page. If on the other hand the jumbled URL also returns 200, we don't
have confidence that the original URL was a useful content page and should
investigate it further.