Clean up list of URLs with jumbled URLs
This file came from a database export which had added NULLs in some empty
fields, and had replaced all 'True' and 'False' values with 0. Clean the
file by using all the values from the CSV file it was based on, and only
keep the values for the four new fields.