Add a quality field to the list of URLs
The comparison of statuses between the original and jumbled URLs and the
searching for 404-like patterns in the page body are taken from @daibach's
local services explorer.

look-like-404-matches are the URLs and matched text which looks like a 404.
I've looked at a sample of these and there are some false positives, eg Ealing
has a homepage title on every page, Test Valley has a feedback form on every
page which says "If you have not found this page helpful...".